+++ src/client/gui/sendmessage/client_sendmessageui.cpp
@@ -150,7 +150,7 @@ mbClientSendMessageUi::mbClientSendMessageUi(QWidget *parent) : mbCoreDialogBase
     m_diagnSubfuncNums.append(MBDIAGN_RETURN_SERVER_NAK_COUNT               );
     m_diagnSubfuncNums.append(MBDIAGN_RETURN_SERVER_BUSY_COUNT              );
     m_diagnSubfuncNums.append(MBDIAGN_RETURN_BUS_CHARACTER_OVERRUN_COUNT    );
-    m_diagnSubfuncNums.append(MBDIAGN_CLEAR_OVERRUN_COUNTER_AND_FLAGS       );
+    m_diagnSubfuncNums.append(MBDIAGN_CLEAR_OVERRUN_COUNTER_AND_FLAG        );
 
     // -----------------------------------------------------------------------
     // Unit